requirements.txt
# ==================================================
# Core Deep Learning (PyTorch ONLY – Python 3.14 safe)
# ==================================================
torch>=2.1.0
torchvision>=0.16.0


# ==================================================
# Computer Vision
# ==================================================
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
ultralytics>=8.1.0
Pillow>=10.2.0


# ==================================================
# Image Processing
# ==================================================
scikit-image>=0.22.0
imageio>=2.34.0
albumentations>=1.4.0


# ==================================================
# Data Science
# ==================================================
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
scikit-learn>=1.4.0


# ==================================================
# Visualization
# ==================================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0


# ==================================================
# Video Processing
# ==================================================
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9


# ==================================================
# Parallel / Performance (Python 3.14 safe)
# ==================================================
joblib>=1.4.0


# ==================================================
# Utilities
# ==================================================
tqdm>=4.66.0
PyYAML>=6.0.1
requests>=2.31.0
gdown>=5.1.0


# ==================================================
# Jupyter / Notebooks
# ==================================================
jupyter>=1.0.0
notebook>=7.1.0
ipywidgets>=8.1.0


# ==================================================
# EXPERIMENT TRACKING (safe)
# ==================================================
wandb>=0.16.0
